fix: use resource file for dummy data to circumvent file location issues
