+++ GameDb.DAL/GameDb.DAL.csproj
@@ -24,4 +24,28 @@
     <ProjectReference Include="..\GameDb.Common\GameDb.Common.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Update="DummyImpl\Resources\Covers.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Covers.resx</DependentUpon>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Update="DummyImpl\Resources\Covers.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Covers.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="DummyImpl\Resources\Covers\11111111-1111-1111-1111-111111111112.png">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="DummyImpl\Resources\Covers\21111111-1111-1111-1111-111111111112.png">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
 </Project>
