+++ GameDb/MauiProgram.cs
@@ -3,6 +3,7 @@
 using CommunityToolkit.Maui;
 
 using Michaelvsk.GameDb.Core.Services;
+using Michaelvsk.GameDb.DAL.DummyImpl;
 using Michaelvsk.GameDb.DAL.Repositories;
 using Michaelvsk.GameDb.Pages;
 using Michaelvsk.GameDb.ViewModels;
